First complete API - Get Students, Get students by id, H2 database, github workflows et resources

+++ src/main/java/fr/efrei/student/server/repository/StudentRepository.java
@@ -1,12 +1,12 @@
-/*
+
 package fr.efrei.student.server.repository;
 
 import fr.efrei.student.server.domain.Student;
 
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface StudentRepository extends JpaRepository <Student, Integer>{
+public interface StudentRepository extends JpaRepository<Student, Long> {
 }
-*/
